+++ homework#1/Taewan/main.go
@@ -1,0 +1,20 @@
+package main
+
+import (
+	"fmt"
+)
+
+//블록체인 동작하는 확인
+func main() {
+	bc := NewBlockchain()
+
+	bc.AddBlock("Send 1 BTC to Ivan")
+	bc.AddBlock("Send 2 more BTC to Ivan")
+
+	for _, block := range bc.blocks {
+		fmt.Printf("Prev. hash: %x\n", block.PrevBlockHash)
+		fmt.Printf("Data: %s\n", block.Data)
+		fmt.Printf("Hash: %x\n", block.Hash)
+		fmt.Println()
+	}
+}
